Opdateret admin-film-edit og admin-film-list... Samt også slettet film-list mappen.

+++ src/app/pages/admin/admin-film-list/admin-film-list.ts
@@ -1,25 +1,31 @@
 import { Component } from '@angular/core';
 import { NgFor } from '@angular/common';
-import { RouterLink } from '@angular/router';  // <-- korrekt
+import { RouterLink } from '@angular/router';
 import { FilmService, Film } from '../../../services/film';
 
 @Component({
   selector: 'app-admin-film-list',
   standalone: true,
-  imports: [NgFor, RouterLink],   // <-- RouterLink SKAL stå her
+  imports: [NgFor, RouterLink],
   templateUrl: './admin-film-list.html',
-  styleUrl: './admin-film-list.css'
+  styleUrl: './admin-film-list.css',
 })
 export class AdminFilmList {
-
   filmListe: Film[] = [];
 
   constructor(private filmService: FilmService) {
+    this.refresh();
+  }
+
+  refresh() {
     this.filmListe = this.filmService.getFilms();
   }
 
   deleteFilm(id: number) {
+    const ok = confirm('Er du sikker på at du vil slette filmen?');
+    if (!ok) return;
+
     this.filmService.deleteFilm(id);
-    this.filmListe = this.filmService.getFilms();
+    this.refresh();
   }
 }
